setting up services directory structure
