+++ phpunit.xml
@@ -4,7 +4,8 @@
     bootstrap="./vendor/autoload.php"
     type="coverage-text" 
     target="src/serviceBundle/" 
-    charset="UTF-8">
+    charset="UTF-8"
+    stopOnFailure="true">
     <testsuites>
       <testsuite name="serviceBundle Test Suite">
         <directory suffix="Test.php">./tests/serviceBundle</directory>
