+++ src/serviceBundle/Service.php
@@ -5,5 +5,5 @@ namespace peter\components\serviceBundle;
 interface Service
 {
     public function sendReq();
-    public function setConfigs($configs);
+    public function setConfig($config);
 }
